map domains to services
In the origin implementation, different ports are mapped to different
services. In this commit, different domains are mapped to different
services.

In addition, keycloak is also back proxied with apache. Now only the
port 9443 need to be exposed, which can be customized in the following
commit.

All the domains can be customized through the corresponding options.